client update 기존코드는 먼저 들어간 유저가 캐릭터 선택 한 뒤에 다른 플레이어가 방에 들어왔을때 바로 첫번째 유저의 캐릭이 렌더링 안되고 두번째 들어온 애가 자신의 캐릭터를 선택한 뒤에야 먼저 들어온 플레이어의 정보를 렌더하는 문제가 있었음.
따라서 먼저 들어간 유저가 직업을 선택했을 때 그 후에 들어온 유저는 들어오면서 이미 들어와 있는 유저들의 정보를 렌더링 하게 바꿈.

+++ Assets/Scripts/Server+Client_Yeram/Lobby/Client_Script/RoomInfoBtn.cs
@@ -23,7 +23,7 @@ public class RoomInfoBtn : MonoBehaviour
     }
     public void OnClick_RoomBtn()
     {
-        LobbyGUIManager.Instance.OnClick_Room(m_roominfo.GetID);
+        LobbyGUIManager.Instance.OnClick_Room((uint)m_roominfo.GetID);
     }
     private void Start()
     {
